requirements.txt
yfinance>=0.2
httpx>=0.27
finnhub-python>=2.4
google-generativeai>=0.8
Jinja2>=3.1
python-dotenv>=1.0
black>=24.0
